requirements.txt
discord.py~=2.1
asyncpg~=0.27.0
APScheduler~=3.9.1.post1